+++ addons/fray/src/input/input_data/binds/input_bind_key.gd
@@ -1,5 +1,5 @@
 tool
-extends "input_bind.gd"
+extends "input_bind_simple.gd"
 ## Keyboard input bind
 
 ## The key scancode, which corresponds to one of the KeyList constants.
@@ -11,3 +11,9 @@ func _init(keyboard_key: int = -1) -> void:
 	
 func _is_pressed_impl(device: int = 0) -> bool:
 	return Input.is_key_pressed(key)
+
+
+func _equals_impl(input_bind: Resource) -> bool:
+	return (
+		._equals_impl(input_bind)
+		and key == input_bind.key)