+++ addons/fray/src/input/input_data/binds/input_bind_mouse_button.gd
@@ -1,5 +1,5 @@
 tool
-extends "input_bind.gd"
+extends "input_bind_simple.gd"
 ## Mouse input bind
 
 ## The mouse button identifier, one of the ButtonList buttons or button wheel constants.
@@ -11,3 +11,9 @@ func _init(mouse_button: int = -1) -> void:
 
 func _is_pressed_impl(device: int = 0) -> bool:
 	return Input.is_mouse_button_pressed(button)
+
+
+func _equals_impl(input_bind: Resource) -> bool:
+	return (
+		._equals_impl(input_bind)
+		and button == input_bind.button)