+++ addons/fray/src/input/input_data/binds/input_bind_simple.gd
@@ -1,0 +1,2 @@
+extends "input_bind.gd"
+## Abstract base class representing all simple inputs